[py] Let setTargetPose fail when invalid kinematic group name passed.
`setTargetPose` takes a kinematic group as its 1st arg. When invalid value passed, it still tries to solve IK and always fails with an IK solution failure message, which is ambiguous because it failed not because the computation but because the target group name is simply invalid.

With this change it first checks if the passed value is a registered kinematic group name and fails with the reason printed.